simplify AssetIdMapping interface (#2091)
* expose union get metadata by assetid

* update type bound

* update type bound
